Additional config to get basic ami and cluster creation working

+++ roles/docker-disk/files/docker-data-volume.sh
@@ -5,6 +5,11 @@
 
 set -e
 
+# If this has already been created assume we've run before and exit
+if [ -e /dev/docker/docker ]; then
+  exit
+fi
+
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
 systemctl stop docker
@@ -12,10 +17,10 @@ yum -y install lvm2
 echo "overlay" >> /etc/modules-load.d/overlay.conf
 modprobe overlay
 mkdir -p /etc/systemd/system/docker.service.d
-rm -rf /var/lib/docker/*
 echo "[Service]
 ExecStart=
 ExecStart=/bin/docker daemon --storage-driver=overlay -H fd://" >> /etc/systemd/system/docker.service.d/override.conf
+systemctl daemon-reload
 
 # Lists all block devices and skips the first two lines containing a column label and /dev/sda boot volume
 DATA_VOLUMES=$(lsblk -d | awk '{ print $1 }' | grep -v "loop" | tail -n+3)
@@ -29,10 +34,15 @@ for v in $DATA_VOLUMES; do
   vgcreate docker /dev/${v}1
   lvcreate --name docker --extents 100%FREE docker
   mkfs.xfs /dev/docker/docker
+  sleep 2
+  if [ ! -d /var/lib/docker ]; then
+     mkdir /var/lib/docker
+  fi
   echo "/dev/docker/docker      /var/lib/docker                 xfs     defaults        0 0" >> /etc/fstab
   mount -a
   break
 done
 
+# pick up the new systemd config files and restart docker
 systemctl start docker
 echo "# Do not delete" > $SCRIPT_DIR/docker-data-volume.done
